# BE/swagger.yaml
openapi: 3.0.0
info:
  title: inventory surat
  contact: {}
  version: '1.0'
servers:
- url: http://localhost:5151/v1/api
  variables: {}
paths:
...
  /upload/{id}:
    post:
      tags:
      - Surat
      summary: upload document
      operationId: uploaddocument
      parameters:
      - name: id
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: integer
          format: int32
          example: 5
      - name: authorization
        in: header
        description: ''
        required: true
        style: simple
        schema:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6NCwicm9sZSI6MSwiaWF0IjoxNjgxMjIzODAyLCJleHAiOjE2ODEzMTAyMDJ9.S5QdmAC-hyoj7veVqfKdDtmApUfZRxk7gE-XcXFyXDE
      requestBody:
        content:
          multipart/form-data:
            encoding: {}
            schema:
              required:
              - file
              type: object
              properties:
                file:
                  type: string
                  format: binary
        required: false
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security: []
...